Refactor: remove AbstractReplicator
This was a bridge-class between open-replicator and binlog-connector.
No need for the abstract functionality any more.
