+++ src/main/java/com/zendesk/maxwell/replication/BinlogConnectorReplicator.java
@@ -1,46 +1,76 @@
 package com.zendesk.maxwell.replication;
 
+import com.codahale.metrics.Counter;
 import com.codahale.metrics.Histogram;
+import com.codahale.metrics.Meter;
 import com.github.shyiko.mysql.binlog.BinaryLogClient;
-import com.github.shyiko.mysql.binlog.event.*;
+import com.github.shyiko.mysql.binlog.event.EventType;
+import com.github.shyiko.mysql.binlog.event.QueryEventData;
+import com.github.shyiko.mysql.binlog.event.RowsQueryEventData;
+import com.github.shyiko.mysql.binlog.event.TableMapEventData;
 import com.github.shyiko.mysql.binlog.event.deserialization.EventDeserializer;
-import com.github.shyiko.mysql.binlog.network.SSLMode;
 import com.zendesk.maxwell.MaxwellContext;
 import com.zendesk.maxwell.MaxwellMysqlConfig;
 import com.zendesk.maxwell.bootstrap.AbstractBootstrapper;
 import com.zendesk.maxwell.filtering.Filter;
 import com.zendesk.maxwell.monitoring.Metrics;
 import com.zendesk.maxwell.producer.AbstractProducer;
+import com.zendesk.maxwell.row.HeartbeatRowMap;
 import com.zendesk.maxwell.row.RowMap;
 import com.zendesk.maxwell.row.RowMapBuffer;
 import com.zendesk.maxwell.schema.Schema;
 import com.zendesk.maxwell.schema.SchemaStore;
 import com.zendesk.maxwell.schema.SchemaStoreException;
 import com.zendesk.maxwell.schema.Table;
+import com.zendesk.maxwell.schema.ddl.DDLMap;
+import com.zendesk.maxwell.schema.ddl.ResolvedSchemaChange;
+import com.zendesk.maxwell.util.RunLoopProcess;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
+import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Pattern;
 
-public class BinlogConnectorReplicator extends AbstractReplicator implements Replicator {
-	private final long MAX_TX_ELEMENTS = 10000;
-	protected SchemaStore schemaStore;
+public class BinlogConnectorReplicator extends RunLoopProcess implements Replicator {
+	static final Logger LOGGER = LoggerFactory.getLogger(BinlogConnectorReplicator.class);
+	private static final long MAX_TX_ELEMENTS = 10000;
 
-	private final LinkedBlockingDeque<BinlogConnectorEvent> queue = new LinkedBlockingDeque<>(20);
-
-	private BinlogConnectorEventListener binlogEventListener;
-	private BinlogConnectorLifecycleListener binlogLifecycleListener;
+	private final String clientID;
+	private final String maxwellSchemaDatabaseName;
 
 	private final BinaryLogClient client;
+	private BinlogConnectorEventListener binlogEventListener;
+	private BinlogConnectorLifecycleListener binlogLifecycleListener;
+	private final LinkedBlockingDeque<BinlogConnectorEvent> queue = new LinkedBlockingDeque<>(20);
+	private final TableCache tableCache = new TableCache();
 
-	static final Logger LOGGER = LoggerFactory.getLogger(BinlogConnectorReplicator.class);
 	private final boolean stopOnEOF;
 	private boolean hitEOF = false;
+
+	private Position lastHeartbeatPosition;
+	private final HeartbeatNotifier heartbeatNotifier;
+	private Long stopAtHeartbeat;
+	private Filter filter;
+
+	private final AbstractBootstrapper bootstrapper;
+	private final AbstractProducer producer;
+	private RowMapBuffer rowBuffer;
+
+	private final Counter rowCounter;
+	private final Meter rowMeter;
+	private SchemaStore schemaStore;
 	private Histogram transactionRowCount;
 	private Histogram transactionExecutionTime;
+
+
+	private static Pattern createTablePattern =
+		Pattern.compile("^CREATE\\s+TABLE", Pattern.CASE_INSENSITIVE);
+
 
 	public BinlogConnectorReplicator(
 		SchemaStore schemaStore,
@@ -55,12 +85,31 @@ public class BinlogConnectorReplicator extends AbstractReplicator implements Rep
 		String clientID,
 		HeartbeatNotifier heartbeatNotifier
 	) {
-		super(clientID, bootstrapper, maxwellSchemaDatabaseName, producer, metrics, start, heartbeatNotifier);
+		this.clientID = clientID;
+		this.bootstrapper = bootstrapper;
+		this.maxwellSchemaDatabaseName = maxwellSchemaDatabaseName;
+		this.producer = producer;
+		this.lastHeartbeatPosition = start;
+		this.heartbeatNotifier = heartbeatNotifier;
+		this.stopOnEOF = stopOnEOF;
 		this.schemaStore = schemaStore;
-		transactionExecutionTime = metrics.getRegistry().histogram(metrics.metricName("transaction", "execution_time"));
-		transactionRowCount = metrics.getRegistry().histogram(metrics.metricName("transaction", "row_count"));
 
+		/* setup metrics */
+		rowCounter = metrics.getRegistry().counter(
+			metrics.metricName("row", "count")
+		);
+
+		rowMeter = metrics.getRegistry().meter(
+			metrics.metricName("row", "meter")
+		);
+
+		transactionRowCount = metrics.getRegistry().histogram(metrics.metricName("transaction", "row_count"));
+		transactionExecutionTime = metrics.getRegistry().histogram(metrics.metricName("transaction", "execution_time"));
+
+		/* setup binlog */
+		this.binlogLifecycleListener = new BinlogConnectorLifecycleListener();
 		this.client = new BinaryLogClient(mysqlConfig.host, mysqlConfig.port, mysqlConfig.user, mysqlConfig.password);
+
 		this.client.setSSLMode(mysqlConfig.sslMode);
 
 		BinlogPosition startBinlog = start.getBinlogPosition();
@@ -82,40 +131,27 @@ public class BinlogConnectorReplicator extends AbstractReplicator implements Rep
 		);
 		this.client.setEventDeserializer(eventDeserializer);
 		this.binlogEventListener = new BinlogConnectorEventListener(client, queue, metrics);
-		this.binlogLifecycleListener = new BinlogConnectorLifecycleListener();
-
 		this.client.setBlocking(!stopOnEOF);
 		this.client.registerEventListener(binlogEventListener);
 		this.client.registerLifecycleListener(binlogLifecycleListener);
 		this.client.setServerId(replicaServerID.intValue());
-
-		this.stopOnEOF = stopOnEOF;
 	}
 
-	public BinlogConnectorReplicator(SchemaStore schemaStore, AbstractProducer producer, AbstractBootstrapper bootstrapper, MaxwellContext ctx, Position start) throws SQLException {
-		this(
-			schemaStore,
-			producer,
-			bootstrapper,
-			ctx.getConfig().replicationMysql,
-			ctx.getConfig().replicaServerID,
-			ctx.getConfig().databaseName,
-			ctx.getMetrics(),
-			start,
-			false,
-			ctx.getConfig().clientID,
-			ctx.getHeartbeatNotifier()
-		);
-	}
+	/**
+	 * get a single row from the replicator and pass it to the producer or bootstrapper.
+	 *
+	 * This is the top-level function in the run-loop.
+	 */
+	public void work() throws Exception {
+		RowMap row = getRow();
 
-	private void ensureReplicatorThread() throws Exception {
-		if ( !client.isConnected() && !stopOnEOF ) {
-			String gtidStr = client.getGtidSet();
-			String binlogPos = client.getBinlogFilename() + ":" + client.getBinlogPosition();
-			String position = gtidStr == null ? binlogPos : gtidStr;
-			LOGGER.warn("replicator stopped at position: " + position + " -- restarting");
-			client.connect(5000);
-		}
+		if ( row == null )
+			return;
+
+		rowCounter.inc();
+		rowMeter.mark();
+
+		processRow(row);
 	}
 
 	public void startReplicator() throws Exception {
@@ -133,8 +169,157 @@ public class BinlogConnectorReplicator extends AbstractReplicator implements Rep
 		this.client.disconnect();
 	}
 
-	private static Pattern createTablePattern =
-			Pattern.compile("^CREATE\\s+TABLE", Pattern.CASE_INSENSITIVE);
+	/**
+	 * Get the last heartbeat that the replicator has processed.
+	 *
+	 * We pass along the value of the heartbeat to the producer inside the row map.
+	 * @return the millisecond value ot the last heartbeat read
+	 */
+
+	public Long getLastHeartbeatRead() {
+		return lastHeartbeatPosition.getLastHeartbeatRead();
+	}
+
+	public void stopAtHeartbeat(long heartbeat) {
+		stopAtHeartbeat = heartbeat;
+	}
+
+	protected void processRow(RowMap row) throws Exception {
+		if ( row instanceof HeartbeatRowMap) {
+			producer.push(row);
+			if (stopAtHeartbeat != null) {
+				long thisHeartbeat = row.getPosition().getLastHeartbeatRead();
+				if (thisHeartbeat >= stopAtHeartbeat) {
+					LOGGER.info("received final heartbeat " + thisHeartbeat + "; stopping replicator");
+					// terminate runLoop
+					this.taskState.stopped();
+				}
+			}
+		} else if (!bootstrapper.shouldSkip(row) && !isMaxwellRow(row))
+			producer.push(row);
+		else
+			bootstrapper.work(row, producer, this);
+	}
+
+
+	/**
+	 * If the input RowMap is one of the heartbeat pulses we sent out,
+	 * process it.  If it's one of our heartbeats, we build a `HeartbeatRowMap`,
+	 * which will be handled specially in producers (namely, it causes the binlog position to advance).
+	 * It is isn't, we leave the row as a RowMap and the rest of the chain will ignore it.
+	 *
+	 * @return either a RowMap or a HeartbeatRowMap
+	 */
+
+	private RowMap processHeartbeats(RowMap row) {
+		String hbClientID = (String) row.getData("client_id");
+		if ( !Objects.equals(hbClientID, this.clientID) )
+			return row; // plain row -- do not process.
+
+		long lastHeartbeatRead = (Long) row.getData("heartbeat");
+		LOGGER.debug("replicator picked up heartbeat: " + lastHeartbeatRead);
+		this.lastHeartbeatPosition = row.getPosition().withHeartbeat(lastHeartbeatRead);
+		heartbeatNotifier.heartbeat(lastHeartbeatRead);
+		return HeartbeatRowMap.valueOf(row.getDatabase(), this.lastHeartbeatPosition, row.getNextPosition().withHeartbeat(lastHeartbeatRead));
+	}
+
+	/**
+	 * Parse a DDL statement and output the results to the producer
+	 *
+	 * @param dbName The database "context" under which the SQL is to be processed.  think "use db; alter table foo ..."
+	 * @param sql The DDL SQL to be processed
+	 * @param schemaStore A SchemaStore object to which we delegate the parsing of the sql
+	 * @param position The position that the SQL happened at
+	 * @param timestamp The timestamp of the SQL binlog event
+	 */
+	private void processQueryEvent(String dbName, String sql, SchemaStore schemaStore, Position position, Position nextPosition, Long timestamp) throws Exception {
+		List<ResolvedSchemaChange> changes = schemaStore.processSQL(sql, dbName, position);
+		for (ResolvedSchemaChange change : changes) {
+			if (change.shouldOutput(filter)) {
+				DDLMap ddl = new DDLMap(change, timestamp, sql, position, nextPosition);
+				producer.push(ddl);
+			}
+		}
+
+		tableCache.clear();
+	}
+
+	private void processQueryEvent(BinlogConnectorEvent event) throws Exception {
+		QueryEventData data = event.queryData();
+		processQueryEvent(
+			data.getDatabase(),
+			data.getSql(),
+			this.schemaStore,
+			Position.valueOf(event.getPosition(), getLastHeartbeatRead()),
+			Position.valueOf(event.getNextPosition(), getLastHeartbeatRead()),
+			event.getEvent().getHeader().getTimestamp()
+		);
+	}
+
+
+	/**
+	 * Should we output a batch of rows for the given database and table?
+	 *
+	 * First against a whitelist/blacklist/filter.  The whitelist
+	 * ensures events that maxwell needs (maxwell.bootstrap, maxwell.heartbeats)
+	 * are always passed along.
+	 *
+	 * The system the blacklist gets rid of the
+	 * `ha_health_check` and `rds_heartbeat` tables which are weird
+	 * replication-control mechanism events in Alibaba RDS (and maybe amazon?)
+	 *
+	 * Then we check the configured filters.
+	 *
+	 * Finall, if we decide to exclude a table we check the filter to
+	 * see if it's possible that a column-value filter could reverse this decision
+	 *
+	 * @param database The database of the DML
+	 * @param table The table of the DML
+	 * @param filter A table-filter, or null
+	 * @param columnNames Names of the columns this table contains
+	 * @return Whether we should write the event to the producer
+	 */
+	private boolean shouldOutputEvent(String database, String table, Filter filter, Set<String> columnNames) {
+		Boolean isSystemWhitelisted = this.maxwellSchemaDatabaseName.equals(database)
+			&& ("bootstrap".equals(table) || "heartbeats".equals(table));
+
+		if ( Filter.isSystemBlacklisted(database, table) )
+			return false;
+		else if ( isSystemWhitelisted )
+			return true;
+		else {
+			if ( Filter.includes(filter, database, table) )
+				return true;
+			else
+				return Filter.couldIncludeFromColumnFilters(filter, database, table, columnNames);
+		}
+	}
+
+
+	private boolean shouldOutputRowMap(String database, String table, RowMap rowMap, Filter filter) {
+		return Filter.includes(filter, database, table, rowMap.getData());
+	}
+
+	/**
+	 * Is this RowMap an update to one of maxwell's own tables?
+	 *
+	 * If so we will often suppress the output.
+	 * @param row The RowMap in question
+	 * @return whether the update is something maxwell itself generated
+	 */
+	protected boolean isMaxwellRow(RowMap row) {
+		return row.getDatabase().equals(this.maxwellSchemaDatabaseName);
+	}
+
+	private void ensureReplicatorThread() throws Exception {
+		if ( !client.isConnected() && !stopOnEOF ) {
+			String gtidStr = client.getGtidSet();
+			String binlogPos = client.getBinlogFilename() + ":" + client.getBinlogPosition();
+			String position = gtidStr == null ? binlogPos : gtidStr;
+			LOGGER.warn("replicator stopped at position: " + position + " -- restarting");
+			client.connect(5000);
+		}
+	}
 
 	/**
 	 * Get a batch of rows for the current transaction.
@@ -227,8 +412,6 @@ public class BinlogConnectorReplicator extends AbstractReplicator implements Rep
 		}
 	}
 
-	private RowMapBuffer rowBuffer;
-
 	/**
 	 * The main entry point into the event reading loop.
 	 *
@@ -314,23 +497,17 @@ public class BinlogConnectorReplicator extends AbstractReplicator implements Rep
 		}
 	}
 
-	protected BinlogConnectorEvent pollEvent() throws InterruptedException {
-		return queue.poll(100, TimeUnit.MILLISECONDS);
+	public void setFilter(Filter filter) {
+		this.filter = filter;
 	}
 
-	private void processQueryEvent(BinlogConnectorEvent event) throws Exception {
-		QueryEventData data = event.queryData();
-		processQueryEvent(
-			data.getDatabase(),
-			data.getSql(),
-			this.schemaStore,
-			Position.valueOf(event.getPosition(), getLastHeartbeatRead()),
-			Position.valueOf(event.getNextPosition(), getLastHeartbeatRead()),
-			event.getEvent().getHeader().getTimestamp()
-		);
+	protected BinlogConnectorEvent pollEvent() throws InterruptedException {
+		return queue.poll(100, TimeUnit.MILLISECONDS);
 	}
 
 	public Schema getSchema() throws SchemaStoreException {
 		return this.schemaStore.getSchema();
 	}
+
+
 }
